fix: Divide InputWithIcon dependencies in components
