+++ components/ui/input.tsx
@@ -2,56 +2,8 @@ import * as React from 'react';
 
 import { cn } from '@/lib/utils';
 
-type IconKey = "user" | "none" | "key";
-
-const icons: Record<IconKey, JSX.Element | string> = {
-  user: (
-    <svg xmlns="http://www.w3.org/2000/svg" fill="currentColor" viewBox="0 0 24 24" className="size-6">
-      <path fillRule="evenodd" d="M7.5 6a4.5 4.5 0 1 1 9 0 4.5 4.5 0 0 1-9 0ZM3.751 20.105a8.25 8.25 0 0 1 16.498 0 .75.75 0 0 1-.437.695A18.683 18.683 0 0 1 12 22.5c-2.786 0-5.433-.608-7.812-1.7a.75.75 0 0 1-.437-.695Z" clipRule="evenodd" />
-    </svg>
-  ),
-  key:<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" fill="currentColor" className="size-6">
-  <path fillRule="evenodd" d="M15.75 1.5a6.75 6.75 0 0 0-6.651 7.906c.067.39-.032.717-.221.906l-6.5 6.499a3 3 0 0 0-.878 2.121v2.818c0 .414.336.75.75.75H6a.75.75 0 0 0 .75-.75v-1.5h1.5A.75.75 0 0 0 9 19.5V18h1.5a.75.75 0 0 0 .53-.22l2.658-2.658c.19-.189.517-.288.906-.22A6.75 6.75 0 1 0 15.75 1.5Zm0 3a.75.75 0 0 0 0 1.5A2.25 2.25 0 0 1 18 8.25a.75.75 0 0 0 1.5 0 3.75 3.75 0 0 0-3.75-3.75Z" clipRule="evenodd" />
-</svg>,
-  none:""
-};
-
 export interface InputProps
   extends React.InputHTMLAttributes<HTMLInputElement> {}
-
-interface InputWithIconProps extends React.InputHTMLAttributes<HTMLInputElement> {
-  icon: IconKey;
-}
-
-const InputWithIcon = React.forwardRef<HTMLInputElement, InputWithIconProps>(
-  ({ className, type, icon = "none", ...props }, ref) => {
-    return (
-      <div
-        className={cn(
-          "group flex items-center gap-1 h-10 w-full rounded-md border border-input bg-background p-2 text-sm ring-offset-background",
-          "focus-within:ring-2 focus-within:ring-ring focus-within:ring-offset-2",
-          "transition-colors",
-        )}
-      >
-        {icon !== "none" && (
-          <div className="w-5 h-5 flex items-center pointer-events-none text-muted-foreground">
-            {icons[icon]}
-          </div>
-        )}
-        <input
-          type={type}
-          className={cn(
-            "flex w-full bg-transparent outline-none text-sm", // <-- bg transparente no input
-            className
-          )}
-          ref={ref}
-          {...props}
-        />
-      </div>
-    );
-  }
-);
-
 
 const Input = React.forwardRef<HTMLInputElement, InputProps>(
   ({ className, type, ...props }, ref) => {
@@ -70,4 +22,4 @@ const Input = React.forwardRef<HTMLInputElement, InputProps>(
 );
 Input.displayName = 'Input';
 
-export { Input, InputWithIcon };
+export { Input };
